Improved the code editor
Now, when you click on the up arrow, when your cursor is on the topmost line, the cursor will move to the beginning. Similarly, when the cursor is at the very bottom: it moves to the very end of the line. Also now, if the cursor is in the middle of the code, and at the same time click on the up/down arrow, the cursor will normally move below, and not to the end of the line, as it was before. Also, now when you click on Tab, the tab is replaced by 4 spaces